feat(main): add books auto download

+++ main.py
@@ -6,147 +6,60 @@ import argparse
 import requests
 import time
 import shutil
-from utils import epub_cfi_converter
-
-colors = {
-    0: "yellow",
-    1: "red",
-    2: "blue",
-    3: "green",
-    4: "violet",
-}
-
-baseBookConfig = {
-    "viewSettings": {
-        "ttsHighlightOptions": {"style": "highlight", "color": "#808080"},
-        "noteExportConfig": {
-            "includeTitle": True,
-            "includeAuthor": True,
-            "includeDate": True,
-            "includeChapterTitles": True,
-            "includeQuotes": True,
-            "includeNotes": True,
-            "includeTimestamp": False,
-            "includeChapterSeparator": False,
-            "noteSeparator": "\n\n",
-            "useCustomTemplate": False,
-            "customTemplate": "",
-            "exportAsPlainText": False,
-        },
-    },
-    "searchConfig": {},
-}
-API_BASE_URL = "https://api.bookmate.ru/api/v5"
-DEFAULT_PER_PAGE = 20
-
-books_ext = (".epub", ".pdf", ".mobi", ".fb2")
-cover_ext = (".png", ".jpg", ".jpeg")
-
-input_directory = Path("input")
-input_directory.mkdir(exist_ok=True)
-processing_directory = Path("processing")
-processing_directory.mkdir(exist_ok=True)
-output_directory = Path("output")
-output_directory.mkdir(exist_ok=True)
-
-
-def md5_file(path, chunk_size=8192):
-    h = hashlib.md5()
-    with open(path, "rb") as f:
-        for chunk in iter(lambda: f.read(chunk_size), b""):
-            h.update(chunk)
-    return h.hexdigest()
-
-
-def get_user_key_data(
-    user_id: str | int, key: str, sleep=0.5, per_page=DEFAULT_PER_PAGE, cache=True
-) -> list:
-    processing_key = processing_directory / f"{key}.json"
-
-    if cache and processing_key.exists():
-        with open(processing_key, "r", encoding="utf-8") as f:
-            return json.load(f)
-
-    page = 1
-    all_books = []
-
-    with requests.Session() as s:
-        while True:
-            time.sleep(sleep)
-
-            params = {"page": page}
-            if per_page != DEFAULT_PER_PAGE:
-                params["per_page"] = per_page
-
-            resp = s.get(
-                f"{API_BASE_URL}/users/{user_id}/{key}",
-                params=params,
-                timeout=10,
-            )
-            resp.raise_for_status()
-
-            data = resp.json()
-            books = data.get(key, [])
-
-            if not books:
-                break
-
-            all_books.extend(books)
-            page += 1
-
-    with open(processing_key, "w", encoding="utf-8") as f:
-        json.dump(all_books, f, ensure_ascii=False)
-
-    return all_books
+from config import input_directory, output_directory, colors, baseBookConfig
+from parser import epub_cfi_converter
+import subprocess
+import sys
+from utils import get_book_data, get_user_key_data
 
 
 def main():
     argparser = argparse.ArgumentParser()
     argparser.add_argument("uuid")
+    argparser.add_argument("--download", action="store_true")
+    argparser.add_argument("--sleep", type=float, default=0.3)
+    argparser.add_argument("--no-cache", dest="use_cache", action="store_false")
     args = argparser.parse_args()
-    USER_ID = args.uuid
-    user_books_uuid = []
-    for book in get_user_key_data(USER_ID, "books"):
-        user_books_uuid.append(book["uuid"])
-    books = []
-    for elem in input_directory.glob("*"):
-        if elem.is_dir():
-            book = {"path": None, "cover": None}
-            for file in elem.iterdir():
-                suffix = file.suffix
-                if suffix in books_ext and book["path"] is None:
-                    book["path"] = file
-                elif suffix in cover_ext and book["cover"] is None:
-                    book["cover"] = file
-                elif suffix == ".json":
-                    json_book = json.loads(file.read_text(encoding="utf-8"))["book"]
-                    book["uuid"] = json_book["uuid"]
-                    if json_book["library_card"]:
-                        lib_card = json_book["library_card"]
-                        book.update(
-                            {"readingStatus": lib_card["state"]}
-                            if lib_card["state"] == "finished"
-                            else {}
-                        )
-                        book.update(
-                            {
-                                "location": (
-                                    f"epubcfi({lib_card['cfi']})"
-                                    if lib_card["cfi"]
-                                    else {}
-                                )
-                            }
-                        )
-                    if len(json_book["publishers"]) > 0:
-                        book["publisher"] = json_book["publishers"][0]["name"]
-                    book["language"] = json_book["language"]
-                    book["title"] = json_book["title"]
-                    book["authors"] = json_book["authors"]
-                    book["description"] = json_book["annotation"]
-            book["hash"] = md5_file(book["path"])
-            books.append(book)
+    USER_ID: str = args.uuid
+    USER_ID = USER_ID if USER_ID.startswith("@") else "@" + USER_ID
+    is_need_download = args.download
+    sleep_time = args.sleep
+    cache = args.use_cache
+    user_books = get_user_key_data(USER_ID, "books", sleep_time, cache=cache)
 
-    quotes = get_user_key_data(USER_ID, "quotes", 0.7, 50)
+    user_books = {book["uuid"]: {} for book in user_books}
+
+    for elem in input_directory.glob("*"):
+        book = get_book_data(elem)
+        if not book:
+            continue
+        user_books[book["uuid"]] = book
+
+    if is_need_download:
+        try:
+            for uuid, book in user_books.items():
+                if not book:
+                    subprocess.run(
+                        [
+                            sys.executable,
+                            "bookmate_downloader/RUBookmatedownloader.py",
+                            "book",
+                            str(uuid),
+                        ]
+                    )
+            books_path = Path("mybooks", "book")
+            for elem in books_path.glob("*"):
+                new_book_path = input_directory / elem.name
+                shutil.move(elem, new_book_path)
+                book = get_book_data(new_book_path)
+                if not book:
+                    continue
+                user_books[book["uuid"]] = book
+        except:
+            pass
+
+    books = [v for _, v in user_books.items() if v]
+    quotes = get_user_key_data(USER_ID, "quotes", sleep_time, 50, cache)
     imported_quotes = {}
     for quote in quotes:
         state = quote["state"]
@@ -211,8 +124,6 @@ def main():
         new_library_books.append(book_card)
         for booknote in book.get("booknotes", []):
             new_note = epub_cfi_converter(book["path"], booknote["cfi"])
-            print(f"{booknote['cfi']} -> {new_note}")
-
             booknote["cfi"] = new_note
         book_config = {
             "updatedAt": current_time,
@@ -220,15 +131,17 @@ def main():
             "booknotes": book.get("booknotes", []),
         }
         if book.get("location"):
-            book_config["location"] = book["location"]
+            location = epub_cfi_converter(book["path"], book["location"])
+            book_config["location"] = location
 
         book_dir = Path(output_directory / book["hash"])
         book_dir.mkdir(exist_ok=True)
         shutil.copy2(book["path"], book_dir)
 
-        cover_src = Path(book["cover"])
-        cover_dst = book_dir / f"cover.png"
-        shutil.copy2(cover_src, cover_dst)
+        if book.get("cover") and Path(book["cover"]).exists():
+            cover_src = Path(book["cover"])
+            cover_dst = book_dir / "cover.png"
+            shutil.copy2(cover_src, cover_dst)
 
         with open(Path(book_dir / "config.json"), "w", encoding="utf-8") as f:
             json.dump(book_config, f, ensure_ascii=False)
